backend/requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
duckduckgo-search
python-dotenv
cors
google-generativeai
